Syncing README.md and ProcBuilderTest.java
Originally the documentation was generated from the test,
however when migrating from googlecode to github and adding
some features I took a couple of shortcuts.

I dug the old tool out (which I should by the way put up on github
and maven central) and massaged the test until the docs where
in alignment again.